Fix failing test

+++ genvm/core/staged_cache_test.go
@@ -20,7 +20,7 @@ func TestCacheGetCopies(t *testing.T) {
 	accountCopy, err := ss.Get(address)
 	require.NoError(t, err)
 	require.Empty(t, accountCopy.Balance)
-	require.Equal(t, 100, account.Balance)
+	require.Equal(t, uint64(100), account.Balance)
 }
 
 func TestCacheUpdatePreserveOrder(t *testing.T) {
